backend/requirements.txt
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9
uuid6>=2024.1.12
fastapi>=0.100
uvicorn>=0.20
pytest>=7.0
alembic>=1.13
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]>=2.0
python-multipart>=0.0.6